+++ jest.config.js
@@ -7,6 +7,19 @@ module.exports = {
     "^.+\\.tsx?$": "ts-jest",
   },
   moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
+  moduleNameMapper: {
+    "^@src/(.*)$": "<rootDir>/src/$1",
+    "^@tests/(.*)$": "<rootDir>/tests/$1",
+    "^@entities/(.*)$": "<rootDir>/src/domain/entities/$1",
+    "^@gateways/(.*)$": "<rootDir>/src/infrastructure/gateways/$1",
+    "^@app-gateways/(.*)$": "<rootDir>/src/application/gateways/$1",
+    "^@interfaces/(.*)$": "<rootDir>/src/interfaces/$1",
+    "^@types/(.*)$": "<rootDir>/src/types/$1",
+    "^@presenters/(.*)$": "<rootDir>/src/presentation/presenters/$1",
+    "^@controllers/(.*)$": "<rootDir>/src/presentation/controllers/$1",
+    "^@presentation-gateways/(.*)$": "<rootDir>/src/presentation/gateways/$1",
+    "^@usecases/(.*)$": "<rootDir>/src/application/use-cases/$1",
+  },
   collectCoverage: true,
   coverageDirectory: "coverage",
   coveragePathIgnorePatterns: ["/node_modules/", "/tests/"],
