+++ package.json
@@ -12,7 +12,7 @@
   "scripts": {
     "dev": "ts-node-dev --respawn --transpile-only -r tsconfig-paths/register src/index.ts",
     "dev:setup": "npx prisma generate && npx prisma db push && npm run db:seed && npm run dev",
-    "build": "tsc",
+    "build": "tsc --project tsconfig.json && tsc-alias -p tsconfig.json",
     "start": "node dist/index.js",
     "test": "jest",
     "test:watch": "jest --watch",
@@ -60,6 +60,7 @@
     "prisma": "^5.10.0",
     "ts-jest": "^29.1.0",
     "ts-node-dev": "^2.0.0",
+    "tsc-alias": "1.8.16",
     "tsconfig-paths": "^4.2.0",
     "typescript": "^5.3.3"
   },
