+++ src/interfaces/repositories/IOrderRepository.ts
@@ -1,4 +1,4 @@
-import Order from "../../domain/entities/Order";
+import Order from "@entities/Order";
 
 export interface IOrderRepository {
   create(order: Order): Promise<Order>;
