+++ src/presentation/controllers/KitchenController.ts
@@ -1,8 +1,8 @@
-import { IDatabaseConnection } from "@src/interfaces/IDbConnection";
+import { IDatabaseConnection } from "@interfaces/IDbConnection";
 import { OrderGateway } from "../gateways/OrderGateway";
-import KitchenUseCases from "../../application/use-cases/KitchenUseCases";
-import OrderPresenter from "../presenters/OrderPresenter";
-import { OrderStatus } from "../../domain/entities/Order";
+import KitchenUseCases from "@usecases/KitchenUseCases";
+import OrderPresenter from "@presenters/OrderPresenter";
+import { OrderStatus } from "@entities/Order";
 
 class KitchenController {
   static async updateOrderStatus(
